moved Product form to a separate file
